// src/lib/types/database.ts
export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          extensions?: Json
          operationName?: string
          query?: string
          variables?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      jobs: {
        Row: {
          application_deadline: string | null
          application_email: string
          business_id: string
          company_name: string
          compensation_amount: string | null
          compensation_type: string
          created_at: string
          description: string
          employment_type: string
          experience_level: string
          id: string
          is_active: boolean
          location: string
          location_type: string
          skills: string[] | null
          title: string
          updated_at: string
          views: number
        }
        Insert: {
          application_deadline?: string | null
          application_email: string
          business_id: string
          company_name: string
          compensation_amount?: string | null
          compensation_type: string
          created_at?: string
          description: string
          employment_type: string
          experience_level: string
          id?: string
          is_active?: boolean
          location: string
          location_type: string
          skills?: string[] | null
          title: string
          updated_at?: string
          views?: number
        }
        Update: {
          application_deadline?: string | null
          application_email?: string
          business_id?: string
          company_name?: string
          compensation_amount?: string | null
          compensation_type?: string
          created_at?: string
          description?: string
          employment_type?: string
          experience_level?: string
          id?: string
          is_active?: boolean
          location?: string
          location_type?: string
          skills?: string[] | null
          title?: string
          updated_at?: string
          views?: number
        }
        Relationships: [
          {
            foreignKeyName: "jobs_business_id_fkey"
            columns: ["business_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          bio: string | null
          company_name: string | null
          created_at: string
          email: string
          first_name: string | null
          github_url: string | null
          id: string
          last_name: string | null
          linkedin_url: string | null
          portfolio_url: string | null
          role: string
          skills: string[] | null
          updated_at: string
        }
        Insert: {
          bio?: string | null
          company_name?: string | null
          created_at?: string
          email: string
          first_name?: string | null
          github_url?: string | null
          id: string
          last_name?: string | null
          linkedin_url?: string | null
          portfolio_url?: string | null
          role: string
          skills?: string[] | null
          updated_at?: string
        }
        Update: {
          bio?: string | null
          company_name?: string | null
          created_at?: string
          email?: string
          first_name?: string | null
          github_url?: string | null
          id?: string
          last_name?: string | null
          linkedin_url?: string | null
          portfolio_url?: string | null
          role?: string
          skills?: string[] | null
          updated_at?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      increment_job_views: { Args: { job_id: string }; Returns: undefined }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never